Add support for a single-file variant for glass
Currently such databases can only be opened for reading - to create
them you need to use xapian-compact (or its API equivalent).

You can embed such databases within another file, and open them by
passing in a file descriptor open on that file and positioned at
the offset the database starts at).

Database::check() is supported.

Fixes #666, reported by Will Greenberg (and previously suggested on
xapian-discuss by Emmanuel Engelhart).

+++ xapian-core/backends/glass/glass_version.cc
@@ -37,6 +37,7 @@
 #include <sys/types.h>
 #include "safesysstat.h"
 #include "safefcntl.h"
+#include "safeunistd.h"
 #include "str.h"
 #include "stringutils.h"
 
@@ -48,8 +49,8 @@
 using namespace std;
 
 /// Glass format version (date of change):
-#define GLASS_FORMAT_VERSION DATE_TO_VERSION(2015,11,12)
-// 2015,11,12 1.3.4 2 bytes "components_of" per item eliminated; more
+#define GLASS_FORMAT_VERSION DATE_TO_VERSION(2015,12,01)
+// 2015,12,01 1.3.4 2 bytes "components_of" per item eliminated; more
 // 2014,11,21 1.3.2 Brass renamed to Glass
 
 /// Convert date <-> version number.  Dates up to 2141-12-31 fit in 2 bytes.
@@ -67,20 +68,51 @@ static const char GLASS_VERSION_MAGIC[GLASS_VERSION_MAGIC_AND_VERSION_LEN] = {
     char((GLASS_FORMAT_VERSION >> 8) & 0xff), char(GLASS_FORMAT_VERSION & 0xff)
 };
 
+GlassVersion::~GlassVersion()
+{
+    // Either this is a single-file database, or this fd is from opening a new
+    // version file in write(), but sync() was never called.
+    if (fd != -1)
+	(void)::close(fd);
+}
+
+void
+GlassVersion::set_fd(int fd_)
+{
+    db_dir = string();
+    fd = fd_;
+    offset = lseek(fd, 0, SEEK_CUR);
+    if (rare(offset == off_t(-1))) {
+	string msg = "lseek failed on file descriptor ";
+	msg += str(fd);
+	throw Xapian::DatabaseOpeningError(msg, errno);
+    }
+}
+
 void
 GlassVersion::read()
 {
     LOGCALL_VOID(DB, "GlassVersion::read", NO_ARGS);
-    string filename = db_dir;
-    filename += "/iamglass";
-    int fd_in = posixy_open(filename.c_str(), O_RDONLY|O_BINARY);
-    if (rare(fd_in < 0)) {
-	string msg = filename;
-	msg += ": Failed to open glass revision file for reading";
-	throw Xapian::DatabaseOpeningError(msg, errno);
+    FD close_fd(-1);
+    int fd_in;
+    if (db_dir.empty()) {
+	if (rare(lseek(fd, offset, SEEK_SET) == off_t(-1))) {
+	    string msg = "Failed to rewind file descriptor ";
+	    msg += str(fd);
+	    throw Xapian::DatabaseOpeningError(msg, errno);
+	}
+	fd_in = fd;
+    } else {
+	string filename = db_dir;
+	filename += "/iamglass";
+	fd_in = posixy_open(filename.c_str(), O_RDONLY|O_BINARY);
+	if (rare(fd_in < 0)) {
+	    string msg = filename;
+	    msg += ": Failed to open glass revision file for reading";
+	    throw Xapian::DatabaseOpeningError(msg, errno);
+	}
+	close_fd = fd_in;
     }
-
-    FD close_fd(fd_in);
 
     char buf[256];
 
@@ -95,8 +127,12 @@ GlassVersion::read()
     version <<= 8;
     version |= static_cast<unsigned char>(buf[GLASS_VERSION_MAGIC_LEN + 1]);
     if (version != GLASS_FORMAT_VERSION) {
-	string msg = filename;
-	msg += ": Database is format version ";
+	string msg;
+	if (!db_dir.empty()) {
+	    msg = db_dir;
+	    msg += ": ";
+	}
+	msg += "Database is format version ";
 	msg += str(VERSION_TO_YEAR(version) * 10000 +
 		   VERSION_TO_MONTH(version) * 100 +
 		   VERSION_TO_DAY(version));
@@ -121,6 +157,9 @@ GlassVersion::read()
 	old_root[table_no] = root[table_no];
     }
 
+    // For a single-file database, this will assign extra data.  We read
+    // sizeof(buf) above, then skip GLASS_VERSION_MAGIC_AND_VERSION_LEN,
+    // then 16, then the size of the serialised root info.
     serialised_stats.assign(p, end);
     unserialise_stats();
 }
@@ -139,10 +178,7 @@ GlassVersion::serialise_stats()
     // be larger than doclen_lbound.
     pack_uint(serialised_stats, doclen_ubound - wdf_ubound);
     pack_uint(serialised_stats, oldest_changeset);
-    // Micro-optimisation: total_doclen is likely to be the largest value, so
-    // store it last as pack_uint_last() uses a slightly more compact encoding
-    // - this could save us a few bytes!
-    pack_uint_last(serialised_stats, total_doclen);
+    pack_uint(serialised_stats, total_doclen);
 }
 
 void
@@ -167,12 +203,17 @@ GlassVersion::unserialise_stats()
 	!unpack_uint(&p, end, &wdf_ubound) ||
 	!unpack_uint(&p, end, &doclen_ubound) ||
 	!unpack_uint(&p, end, &oldest_changeset) ||
-	!unpack_uint_last(&p, end, &total_doclen)) {
+	!unpack_uint(&p, end, &total_doclen)) {
 	const char * m = p ?
 	    "Bad serialised DB stats (overflowed)" :
 	    "Bad serialised DB stats (out of data)";
 	throw Xapian::DatabaseCorruptError(m);
     }
+
+    // In the single-file DB case, there will be extra data in
+    // serialised_stats, so suppress this check.
+    if (p != end && !db_dir.empty())
+	throw Xapian::DatabaseCorruptError("Rev file has junk at end");
 
     // last_docid must always be >= doccount.
     last_docid += doccount;
@@ -232,22 +273,29 @@ GlassVersion::write(glass_revision_number_t new_rev, int flags)
     serialise_stats();
     s += serialised_stats;
 
-    string tmpfile = db_dir;
-    // In dangerous mode, just write the new version file in place.
-    if (flags & Xapian::DB_DANGEROUS)
-	tmpfile += "/iamglass";
-    else
-	tmpfile += "/v.tmp";
+    string tmpfile;
+    if (!db_dir.empty()) {
+	tmpfile = db_dir;
+	// In dangerous mode, just write the new version file in place.
+	if (flags & Xapian::DB_DANGEROUS)
+	    tmpfile += "/iamglass";
+	else
+	    tmpfile += "/v.tmp";
 
-    fd = posixy_open(tmpfile.c_str(), O_CREAT|O_TRUNC|O_WRONLY|O_BINARY, 0666);
-    if (rare(fd < 0))
-	throw Xapian::DatabaseOpeningError("Couldn't write new rev file: " + tmpfile,
-					   errno);
+	fd = posixy_open(tmpfile.c_str(), O_CREAT|O_TRUNC|O_WRONLY|O_BINARY, 0666);
+	if (rare(fd < 0))
+	    throw Xapian::DatabaseOpeningError("Couldn't write new rev file: " + tmpfile,
+					       errno);
+
+	if (flags & Xapian::DB_DANGEROUS)
+	    tmpfile = string();
+    }
 
     try {
 	io_write(fd, s.data(), s.size());
     } catch (...) {
-	(void)close(fd);
+	if (!db_dir.empty())
+	    (void)close(fd);
 	throw;
     }
 
@@ -260,8 +308,6 @@ GlassVersion::write(glass_revision_number_t new_rev, int flags)
 	changes->write_block(s);
     }
 
-    if (flags & Xapian::DB_DANGEROUS)
-	tmpfile = string();
     RETURN(tmpfile);
 }
 
@@ -271,39 +317,50 @@ GlassVersion::sync(const string & tmpfile,
 {
     Assert(new_rev > rev || rev == 0);
 
-    if ((flags & Xapian::DB_NO_SYNC) == 0 &&
-	((flags & Xapian::DB_FULL_SYNC) ?
-	  !io_full_sync(fd) :
-	  !io_sync(fd))) {
-	int save_errno = errno;
-	(void)close(fd);
-	if (!tmpfile.empty())
-	    (void)unlink(tmpfile.c_str());
-	errno = save_errno;
-	return false;
-    }
-
-    if (close(fd) != 0) {
-	if (!tmpfile.empty()) {
-	    int save_errno = errno;
-	    (void)unlink(tmpfile.c_str());
-	    errno = save_errno;
+    if (db_dir.empty()) {
+	if ((flags & Xapian::DB_NO_SYNC) == 0 &&
+	    ((flags & Xapian::DB_FULL_SYNC) ?
+	      !io_full_sync(fd) :
+	      !io_sync(fd))) {
+	    // FIXME what to do?
 	}
-	return false;
-    }
-
-    if (!tmpfile.empty()) {
-	string filename = db_dir;
-	filename += "/iamglass";
-
-	if (posixy_rename(tmpfile.c_str(), filename.c_str()) < 0) {
-	    // Over NFS, rename() can sometimes report failure when the
-	    // operation succeeded, so in this case we try to unlink the source
-	    // to check if the rename really failed.
+    } else {
+	int fd_to_close = fd;
+	fd = -1;
+	if ((flags & Xapian::DB_NO_SYNC) == 0 &&
+	    ((flags & Xapian::DB_FULL_SYNC) ?
+	      !io_full_sync(fd_to_close) :
+	      !io_sync(fd_to_close))) {
 	    int save_errno = errno;
-	    if (unlink(tmpfile.c_str()) == 0 || errno != ENOENT) {
+	    (void)close(fd_to_close);
+	    if (!tmpfile.empty())
+		(void)unlink(tmpfile.c_str());
+	    errno = save_errno;
+	    return false;
+	}
+
+	if (close(fd_to_close) != 0) {
+	    if (!tmpfile.empty()) {
+		int save_errno = errno;
+		(void)unlink(tmpfile.c_str());
 		errno = save_errno;
-		return false;
+	    }
+	    return false;
+	}
+
+	if (!tmpfile.empty()) {
+	    string filename = db_dir;
+	    filename += "/iamglass";
+
+	    if (posixy_rename(tmpfile.c_str(), filename.c_str()) < 0) {
+		// Over NFS, rename() can sometimes report failure when the
+		// operation succeeded, so in this case we try to unlink the source
+		// to check if the rename really failed.
+		int save_errno = errno;
+		if (unlink(tmpfile.c_str()) == 0 || errno != ENOENT) {
+		    errno = save_errno;
+		    return false;
+		}
 	    }
 	}
     }
